requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
